requirements.txt
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.2
joblib==1.3.2
pyyaml==6.0.1
imbalanced-learn==0.11.0